+++ spec/unit/braintree/transaction_spec.rb
@@ -171,25 +171,6 @@ describe Braintree::Transaction do
       transaction.three_d_secure_info.liability_shift_possible.should == true
     end
 
-    it "sets up ideal_payment_details" do
-      transaction = Braintree::Transaction._new(
-        :gateway,
-        :ideal_payment => {
-          :ideal_payment_id => "idealpayment_abc_123",
-          :ideal_transaction_id => "1150000008857321",
-          :masked_iban => "12************7890",
-          :bic => "RABONL2U",
-          :image_url => "http://www.example.com/ideal.png"
-        }
-      )
-
-      transaction.ideal_payment_details.ideal_payment_id.should == "idealpayment_abc_123"
-      transaction.ideal_payment_details.ideal_transaction_id.should == "1150000008857321"
-      transaction.ideal_payment_details.masked_iban.should == "12************7890"
-      transaction.ideal_payment_details.bic.should == "RABONL2U"
-      transaction.ideal_payment_details.image_url.should == "http://www.example.com/ideal.png"
-    end
-
     it "sets up history attributes in status_history" do
       time = Time.utc(2010,1,14)
       transaction = Braintree::Transaction._new(